TST: Timestamp tests compiled (gh14854)
