+++ pandas/tseries/tests/test_timeseries.py
@@ -2947,423 +2947,6 @@ class TestSeriesDatetime64(tm.TestCase):
                 assert_series_equal(round_method(freq), expected)
 
 
-class TestTimestamp(tm.TestCase):
-    def test_class_ops_pytz(self):
-        tm._skip_if_no_pytz()
-        from pytz import timezone
-
-        def compare(x, y):
-            self.assertEqual(int(Timestamp(x).value / 1e9),
-                             int(Timestamp(y).value / 1e9))
-
-        compare(Timestamp.now(), datetime.now())
-        compare(Timestamp.now('UTC'), datetime.now(timezone('UTC')))
-        compare(Timestamp.utcnow(), datetime.utcnow())
-        compare(Timestamp.today(), datetime.today())
-        current_time = calendar.timegm(datetime.now().utctimetuple())
-        compare(Timestamp.utcfromtimestamp(current_time),
-                datetime.utcfromtimestamp(current_time))
-        compare(Timestamp.fromtimestamp(current_time),
-                datetime.fromtimestamp(current_time))
-
-        date_component = datetime.utcnow()
-        time_component = (date_component + timedelta(minutes=10)).time()
-        compare(Timestamp.combine(date_component, time_component),
-                datetime.combine(date_component, time_component))
-
-    def test_class_ops_dateutil(self):
-        tm._skip_if_no_dateutil()
-        from dateutil.tz import tzutc
-
-        def compare(x, y):
-            self.assertEqual(int(np.round(Timestamp(x).value / 1e9)),
-                             int(np.round(Timestamp(y).value / 1e9)))
-
-        compare(Timestamp.now(), datetime.now())
-        compare(Timestamp.now('UTC'), datetime.now(tzutc()))
-        compare(Timestamp.utcnow(), datetime.utcnow())
-        compare(Timestamp.today(), datetime.today())
-        current_time = calendar.timegm(datetime.now().utctimetuple())
-        compare(Timestamp.utcfromtimestamp(current_time),
-                datetime.utcfromtimestamp(current_time))
-        compare(Timestamp.fromtimestamp(current_time),
-                datetime.fromtimestamp(current_time))
-
-        date_component = datetime.utcnow()
-        time_component = (date_component + timedelta(minutes=10)).time()
-        compare(Timestamp.combine(date_component, time_component),
-                datetime.combine(date_component, time_component))
-
-    def test_basics_nanos(self):
-        val = np.int64(946684800000000000).view('M8[ns]')
-        stamp = Timestamp(val.view('i8') + 500)
-        self.assertEqual(stamp.year, 2000)
-        self.assertEqual(stamp.month, 1)
-        self.assertEqual(stamp.microsecond, 0)
-        self.assertEqual(stamp.nanosecond, 500)
-
-        # GH 14415
-        val = np.iinfo(np.int64).min + 80000000000000
-        stamp = Timestamp(val)
-        self.assertEqual(stamp.year, 1677)
-        self.assertEqual(stamp.month, 9)
-        self.assertEqual(stamp.day, 21)
-        self.assertEqual(stamp.microsecond, 145224)
-        self.assertEqual(stamp.nanosecond, 192)
-
-    def test_unit(self):
-
-        def check(val, unit=None, h=1, s=1, us=0):
-            stamp = Timestamp(val, unit=unit)
-            self.assertEqual(stamp.year, 2000)
-            self.assertEqual(stamp.month, 1)
-            self.assertEqual(stamp.day, 1)
-            self.assertEqual(stamp.hour, h)
-            if unit != 'D':
-                self.assertEqual(stamp.minute, 1)
-                self.assertEqual(stamp.second, s)
-                self.assertEqual(stamp.microsecond, us)
-            else:
-                self.assertEqual(stamp.minute, 0)
-                self.assertEqual(stamp.second, 0)
-                self.assertEqual(stamp.microsecond, 0)
-            self.assertEqual(stamp.nanosecond, 0)
-
-        ts = Timestamp('20000101 01:01:01')
-        val = ts.value
-        days = (ts - Timestamp('1970-01-01')).days
-
-        check(val)
-        check(val / long(1000), unit='us')
-        check(val / long(1000000), unit='ms')
-        check(val / long(1000000000), unit='s')
-        check(days, unit='D', h=0)
-
-        # using truediv, so these are like floats
-        if compat.PY3:
-            check((val + 500000) / long(1000000000), unit='s', us=500)
-            check((val + 500000000) / long(1000000000), unit='s', us=500000)
-            check((val + 500000) / long(1000000), unit='ms', us=500)
-
-        # get chopped in py2
-        else:
-            check((val + 500000) / long(1000000000), unit='s')
-            check((val + 500000000) / long(1000000000), unit='s')
-            check((val + 500000) / long(1000000), unit='ms')
-
-        # ok
-        check((val + 500000) / long(1000), unit='us', us=500)
-        check((val + 500000000) / long(1000000), unit='ms', us=500000)
-
-        # floats
-        check(val / 1000.0 + 5, unit='us', us=5)
-        check(val / 1000.0 + 5000, unit='us', us=5000)
-        check(val / 1000000.0 + 0.5, unit='ms', us=500)
-        check(val / 1000000.0 + 0.005, unit='ms', us=5)
-        check(val / 1000000000.0 + 0.5, unit='s', us=500000)
-        check(days + 0.5, unit='D', h=12)
-
-        # nan
-        result = Timestamp(np.nan)
-        self.assertIs(result, NaT)
-
-        result = Timestamp(None)
-        self.assertIs(result, NaT)
-
-        result = Timestamp(iNaT)
-        self.assertIs(result, NaT)
-
-        result = Timestamp(NaT)
-        self.assertIs(result, NaT)
-
-        result = Timestamp('NaT')
-        self.assertIs(result, NaT)
-
-        self.assertTrue(isnull(Timestamp('nat')))
-
-    def test_roundtrip(self):
-
-        # test value to string and back conversions
-        # further test accessors
-        base = Timestamp('20140101 00:00:00')
-
-        result = Timestamp(base.value + pd.Timedelta('5ms').value)
-        self.assertEqual(result, Timestamp(str(base) + ".005000"))
-        self.assertEqual(result.microsecond, 5000)
-
-        result = Timestamp(base.value + pd.Timedelta('5us').value)
-        self.assertEqual(result, Timestamp(str(base) + ".000005"))
-        self.assertEqual(result.microsecond, 5)
-
-        result = Timestamp(base.value + pd.Timedelta('5ns').value)
-        self.assertEqual(result, Timestamp(str(base) + ".000000005"))
-        self.assertEqual(result.nanosecond, 5)
-        self.assertEqual(result.microsecond, 0)
-
-        result = Timestamp(base.value + pd.Timedelta('6ms 5us').value)
-        self.assertEqual(result, Timestamp(str(base) + ".006005"))
-        self.assertEqual(result.microsecond, 5 + 6 * 1000)
-
-        result = Timestamp(base.value + pd.Timedelta('200ms 5us').value)
-        self.assertEqual(result, Timestamp(str(base) + ".200005"))
-        self.assertEqual(result.microsecond, 5 + 200 * 1000)
-
-    def test_comparison(self):
-        # 5-18-2012 00:00:00.000
-        stamp = long(1337299200000000000)
-
-        val = Timestamp(stamp)
-
-        self.assertEqual(val, val)
-        self.assertFalse(val != val)
-        self.assertFalse(val < val)
-        self.assertTrue(val <= val)
-        self.assertFalse(val > val)
-        self.assertTrue(val >= val)
-
-        other = datetime(2012, 5, 18)
-        self.assertEqual(val, other)
-        self.assertFalse(val != other)
-        self.assertFalse(val < other)
-        self.assertTrue(val <= other)
-        self.assertFalse(val > other)
-        self.assertTrue(val >= other)
-
-        other = Timestamp(stamp + 100)
-
-        self.assertNotEqual(val, other)
-        self.assertNotEqual(val, other)
-        self.assertTrue(val < other)
-        self.assertTrue(val <= other)
-        self.assertTrue(other > val)
-        self.assertTrue(other >= val)
-
-    def test_compare_invalid(self):
-
-        # GH 8058
-        val = Timestamp('20130101 12:01:02')
-        self.assertFalse(val == 'foo')
-        self.assertFalse(val == 10.0)
-        self.assertFalse(val == 1)
-        self.assertFalse(val == long(1))
-        self.assertFalse(val == [])
-        self.assertFalse(val == {'foo': 1})
-        self.assertFalse(val == np.float64(1))
-        self.assertFalse(val == np.int64(1))
-
-        self.assertTrue(val != 'foo')
-        self.assertTrue(val != 10.0)
-        self.assertTrue(val != 1)
-        self.assertTrue(val != long(1))
-        self.assertTrue(val != [])
-        self.assertTrue(val != {'foo': 1})
-        self.assertTrue(val != np.float64(1))
-        self.assertTrue(val != np.int64(1))
-
-        # ops testing
-        df = DataFrame(randn(5, 2))
-        a = df[0]
-        b = Series(randn(5))
-        b.name = Timestamp('2000-01-01')
-        tm.assert_series_equal(a / b, 1 / (b / a))
-
-    def test_cant_compare_tz_naive_w_aware(self):
-        tm._skip_if_no_pytz()
-        # #1404
-        a = Timestamp('3/12/2012')
-        b = Timestamp('3/12/2012', tz='utc')
-
-        self.assertRaises(Exception, a.__eq__, b)
-        self.assertRaises(Exception, a.__ne__, b)
-        self.assertRaises(Exception, a.__lt__, b)
-        self.assertRaises(Exception, a.__gt__, b)
-        self.assertRaises(Exception, b.__eq__, a)
-        self.assertRaises(Exception, b.__ne__, a)
-        self.assertRaises(Exception, b.__lt__, a)
-        self.assertRaises(Exception, b.__gt__, a)
-
-        if sys.version_info < (3, 3):
-            self.assertRaises(Exception, a.__eq__, b.to_pydatetime())
-            self.assertRaises(Exception, a.to_pydatetime().__eq__, b)
-        else:
-            self.assertFalse(a == b.to_pydatetime())
-            self.assertFalse(a.to_pydatetime() == b)
-
-    def test_cant_compare_tz_naive_w_aware_explicit_pytz(self):
-        tm._skip_if_no_pytz()
-        from pytz import utc
-        # #1404
-        a = Timestamp('3/12/2012')
-        b = Timestamp('3/12/2012', tz=utc)
-
-        self.assertRaises(Exception, a.__eq__, b)
-        self.assertRaises(Exception, a.__ne__, b)
-        self.assertRaises(Exception, a.__lt__, b)
-        self.assertRaises(Exception, a.__gt__, b)
-        self.assertRaises(Exception, b.__eq__, a)
-        self.assertRaises(Exception, b.__ne__, a)
-        self.assertRaises(Exception, b.__lt__, a)
-        self.assertRaises(Exception, b.__gt__, a)
-
-        if sys.version_info < (3, 3):
-            self.assertRaises(Exception, a.__eq__, b.to_pydatetime())
-            self.assertRaises(Exception, a.to_pydatetime().__eq__, b)
-        else:
-            self.assertFalse(a == b.to_pydatetime())
-            self.assertFalse(a.to_pydatetime() == b)
-
-    def test_cant_compare_tz_naive_w_aware_dateutil(self):
-        tm._skip_if_no_dateutil()
-        from dateutil.tz import tzutc
-        utc = tzutc()
-        # #1404
-        a = Timestamp('3/12/2012')
-        b = Timestamp('3/12/2012', tz=utc)
-
-        self.assertRaises(Exception, a.__eq__, b)
-        self.assertRaises(Exception, a.__ne__, b)
-        self.assertRaises(Exception, a.__lt__, b)
-        self.assertRaises(Exception, a.__gt__, b)
-        self.assertRaises(Exception, b.__eq__, a)
-        self.assertRaises(Exception, b.__ne__, a)
-        self.assertRaises(Exception, b.__lt__, a)
-        self.assertRaises(Exception, b.__gt__, a)
-
-        if sys.version_info < (3, 3):
-            self.assertRaises(Exception, a.__eq__, b.to_pydatetime())
-            self.assertRaises(Exception, a.to_pydatetime().__eq__, b)
-        else:
-            self.assertFalse(a == b.to_pydatetime())
-            self.assertFalse(a.to_pydatetime() == b)
-
-    def test_delta_preserve_nanos(self):
-        val = Timestamp(long(1337299200000000123))
-        result = val + timedelta(1)
-        self.assertEqual(result.nanosecond, val.nanosecond)
-
-    def test_frequency_misc(self):
-        self.assertEqual(frequencies.get_freq_group('T'),
-                         frequencies.FreqGroup.FR_MIN)
-
-        code, stride = frequencies.get_freq_code(offsets.Hour())
-        self.assertEqual(code, frequencies.FreqGroup.FR_HR)
-
-        code, stride = frequencies.get_freq_code((5, 'T'))
-        self.assertEqual(code, frequencies.FreqGroup.FR_MIN)
-        self.assertEqual(stride, 5)
-
-        offset = offsets.Hour()
-        result = frequencies.to_offset(offset)
-        self.assertEqual(result, offset)
-
-        result = frequencies.to_offset((5, 'T'))
-        expected = offsets.Minute(5)
-        self.assertEqual(result, expected)
-
-        self.assertRaises(ValueError, frequencies.get_freq_code, (5, 'baz'))
-
-        self.assertRaises(ValueError, frequencies.to_offset, '100foo')
-
-        self.assertRaises(ValueError, frequencies.to_offset, ('', ''))
-
-        with tm.assert_produces_warning(FutureWarning, check_stacklevel=False):
-            result = frequencies.get_standard_freq(offsets.Hour())
-        self.assertEqual(result, 'H')
-
-    def test_hash_equivalent(self):
-        d = {datetime(2011, 1, 1): 5}
-        stamp = Timestamp(datetime(2011, 1, 1))
-        self.assertEqual(d[stamp], 5)
-
-    def test_timestamp_compare_scalars(self):
-        # case where ndim == 0
-        lhs = np.datetime64(datetime(2013, 12, 6))
-        rhs = Timestamp('now')
-        nat = Timestamp('nat')
-
-        ops = {'gt': 'lt',
-               'lt': 'gt',
-               'ge': 'le',
-               'le': 'ge',
-               'eq': 'eq',
-               'ne': 'ne'}
-
-        for left, right in ops.items():
-            left_f = getattr(operator, left)
-            right_f = getattr(operator, right)
-            expected = left_f(lhs, rhs)
-
-            result = right_f(rhs, lhs)
-            self.assertEqual(result, expected)
-
-            expected = left_f(rhs, nat)
-            result = right_f(nat, rhs)
-            self.assertEqual(result, expected)
-
-    def test_timestamp_compare_series(self):
-        # make sure we can compare Timestamps on the right AND left hand side
-        # GH4982
-        s = Series(date_range('20010101', periods=10), name='dates')
-        s_nat = s.copy(deep=True)
-
-        s[0] = pd.Timestamp('nat')
-        s[3] = pd.Timestamp('nat')
-
-        ops = {'lt': 'gt', 'le': 'ge', 'eq': 'eq', 'ne': 'ne'}
-
-        for left, right in ops.items():
-            left_f = getattr(operator, left)
-            right_f = getattr(operator, right)
-
-            # no nats
-            expected = left_f(s, Timestamp('20010109'))
-            result = right_f(Timestamp('20010109'), s)
-            tm.assert_series_equal(result, expected)
-
-            # nats
-            expected = left_f(s, Timestamp('nat'))
-            result = right_f(Timestamp('nat'), s)
-            tm.assert_series_equal(result, expected)
-
-            # compare to timestamp with series containing nats
-            expected = left_f(s_nat, Timestamp('20010109'))
-            result = right_f(Timestamp('20010109'), s_nat)
-            tm.assert_series_equal(result, expected)
-
-            # compare to nat with series containing nats
-            expected = left_f(s_nat, Timestamp('nat'))
-            result = right_f(Timestamp('nat'), s_nat)
-            tm.assert_series_equal(result, expected)
-
-    def test_is_leap_year(self):
-        # GH 13727
-        for tz in [None, 'UTC', 'US/Eastern', 'Asia/Tokyo']:
-            dt = Timestamp('2000-01-01 00:00:00', tz=tz)
-            self.assertTrue(dt.is_leap_year)
-            self.assertIsInstance(dt.is_leap_year, bool)
-
-            dt = Timestamp('1999-01-01 00:00:00', tz=tz)
-            self.assertFalse(dt.is_leap_year)
-
-            dt = Timestamp('2004-01-01 00:00:00', tz=tz)
-            self.assertTrue(dt.is_leap_year)
-
-            dt = Timestamp('2100-01-01 00:00:00', tz=tz)
-            self.assertFalse(dt.is_leap_year)
-
-        self.assertFalse(pd.NaT.is_leap_year)
-        self.assertIsInstance(pd.NaT.is_leap_year, bool)
-
-    def test_round_nat(self):
-        # GH14940
-        ts = Timestamp('nat')
-        print(dir(ts))
-        for method in ["round", "floor", "ceil"]:
-            round_method = getattr(ts, method)
-            for freq in ["s", "5s", "min", "5min", "h", "5h"]:
-                self.assertIs(round_method(freq), ts)
-
-
 class TestSlicing(tm.TestCase):
     def test_slice_year(self):
         dti = DatetimeIndex(freq='B', start=datetime(2005, 1, 1), periods=500)
@@ -4078,73 +3661,6 @@ class TestGuessDatetimeFormat(tm.TestCase):
             np.array(
                 [np.nan, np.nan, np.nan], dtype='O'))
         self.assertTrue(format_for_string_of_nans is None)
-
-
-class TestTimestampToJulianDate(tm.TestCase):
-    def test_compare_1700(self):
-        r = Timestamp('1700-06-23').to_julian_date()
-        self.assertEqual(r, 2342145.5)
-
-    def test_compare_2000(self):
-        r = Timestamp('2000-04-12').to_julian_date()
-        self.assertEqual(r, 2451646.5)
-
-    def test_compare_2100(self):
-        r = Timestamp('2100-08-12').to_julian_date()
-        self.assertEqual(r, 2488292.5)
-
-    def test_compare_hour01(self):
-        r = Timestamp('2000-08-12T01:00:00').to_julian_date()
-        self.assertEqual(r, 2451768.5416666666666666)
-
-    def test_compare_hour13(self):
-        r = Timestamp('2000-08-12T13:00:00').to_julian_date()
-        self.assertEqual(r, 2451769.0416666666666666)
-
-
-class TestDateTimeIndexToJulianDate(tm.TestCase):
-    def test_1700(self):
-        r1 = Float64Index([2345897.5, 2345898.5, 2345899.5, 2345900.5,
-                           2345901.5])
-        r2 = date_range(start=Timestamp('1710-10-01'), periods=5,
-                        freq='D').to_julian_date()
-        self.assertIsInstance(r2, Float64Index)
-        tm.assert_index_equal(r1, r2)
-
-    def test_2000(self):
-        r1 = Float64Index([2451601.5, 2451602.5, 2451603.5, 2451604.5,
-                           2451605.5])
-        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
-                        freq='D').to_julian_date()
-        self.assertIsInstance(r2, Float64Index)
-        tm.assert_index_equal(r1, r2)
-
-    def test_hour(self):
-        r1 = Float64Index(
-            [2451601.5, 2451601.5416666666666666, 2451601.5833333333333333,
-             2451601.625, 2451601.6666666666666666])
-        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
-                        freq='H').to_julian_date()
-        self.assertIsInstance(r2, Float64Index)
-        tm.assert_index_equal(r1, r2)
-
-    def test_minute(self):
-        r1 = Float64Index(
-            [2451601.5, 2451601.5006944444444444, 2451601.5013888888888888,
-             2451601.5020833333333333, 2451601.5027777777777777])
-        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
-                        freq='T').to_julian_date()
-        self.assertIsInstance(r2, Float64Index)
-        tm.assert_index_equal(r1, r2)
-
-    def test_second(self):
-        r1 = Float64Index(
-            [2451601.5, 2451601.500011574074074, 2451601.5000231481481481,
-             2451601.5000347222222222, 2451601.5000462962962962])
-        r2 = date_range(start=Timestamp('2000-02-27'), periods=5,
-                        freq='S').to_julian_date()
-        self.assertIsInstance(r2, Float64Index)
-        tm.assert_index_equal(r1, r2)
 
 
 class TestDaysInMonth(tm.TestCase):
