+++ pandas/tseries/tests/test_tslib.py
@@ -1,29 +1,17 @@
-import nose
-from distutils.version import LooseVersion
-import numpy as np
-
-from pandas import tslib, lib
-import pandas._period as period
 import datetime
+import numpy as np
+from distutils.version import LooseVersion
 
 import pandas as pd
-from pandas.core.api import (Timestamp, Index, Series, Timedelta, Period,
-                             to_datetime)
-from pandas.tslib import get_timezone
-from pandas._period import period_asfreq, period_ordinal
-from pandas.tseries.index import date_range, DatetimeIndex
-from pandas.tseries.frequencies import (
-    get_freq,
-    RESO_US, RESO_MS, RESO_SEC, RESO_HR, RESO_DAY, RESO_MIN
-)
-import pandas.tseries.tools as tools
-import pandas.tseries.offsets as offsets
 import pandas.util.testing as tm
-import pandas.compat as compat
-from pandas.compat.numpy import (np_datetime64_compat,
-                                 np_array_datetime64_compat)
-
-from pandas.util.testing import assert_series_equal, _skip_if_has_locale
+from pandas import tslib, lib, compat
+from pandas.tseries import offsets, tools
+from pandas.tseries.frequencies import get_freq
+from pandas.tseries.index import date_range, DatetimeIndex
+from pandas.util.testing import _skip_if_has_locale
+from pandas._period import period_ordinal, period_asfreq
+from pandas.compat.numpy import np_array_datetime64_compat
+from pandas.core.api import Timestamp, to_datetime, Index, Series
 
 
 class TestTsUtil(tm.TestCase):
@@ -58,589 +46,6 @@ class TestTsUtil(tm.TestCase):
             self.assertEqual(
                 Timestamp(Timestamp.min.to_pydatetime()).value / 1000,
                 Timestamp.min.value / 1000)
-
-
-class TestTimestamp(tm.TestCase):
-
-    def test_constructor(self):
-        base_str = '2014-07-01 09:00'
-        base_dt = datetime.datetime(2014, 7, 1, 9)
-        base_expected = 1404205200000000000
-
-        # confirm base representation is correct
-        import calendar
-        self.assertEqual(calendar.timegm(base_dt.timetuple()) * 1000000000,
-                         base_expected)
-
-        tests = [(base_str, base_dt, base_expected),
-                 ('2014-07-01 10:00', datetime.datetime(2014, 7, 1, 10),
-                  base_expected + 3600 * 1000000000),
-                 ('2014-07-01 09:00:00.000008000',
-                  datetime.datetime(2014, 7, 1, 9, 0, 0, 8),
-                  base_expected + 8000),
-                 ('2014-07-01 09:00:00.000000005',
-                  Timestamp('2014-07-01 09:00:00.000000005'),
-                  base_expected + 5)]
-
-        tm._skip_if_no_pytz()
-        tm._skip_if_no_dateutil()
-        import pytz
-        import dateutil
-        timezones = [(None, 0), ('UTC', 0), (pytz.utc, 0), ('Asia/Tokyo', 9),
-                     ('US/Eastern', -4), ('dateutil/US/Pacific', -7),
-                     (pytz.FixedOffset(-180), -3),
-                     (dateutil.tz.tzoffset(None, 18000), 5)]
-
-        for date_str, date, expected in tests:
-            for result in [Timestamp(date_str), Timestamp(date)]:
-                # only with timestring
-                self.assertEqual(result.value, expected)
-                self.assertEqual(tslib.pydt_to_i8(result), expected)
-
-                # re-creation shouldn't affect to internal value
-                result = Timestamp(result)
-                self.assertEqual(result.value, expected)
-                self.assertEqual(tslib.pydt_to_i8(result), expected)
-
-            # with timezone
-            for tz, offset in timezones:
-                for result in [Timestamp(date_str, tz=tz), Timestamp(date,
-                                                                     tz=tz)]:
-                    expected_tz = expected - offset * 3600 * 1000000000
-                    self.assertEqual(result.value, expected_tz)
-                    self.assertEqual(tslib.pydt_to_i8(result), expected_tz)
-
-                    # should preserve tz
-                    result = Timestamp(result)
-                    self.assertEqual(result.value, expected_tz)
-                    self.assertEqual(tslib.pydt_to_i8(result), expected_tz)
-
-                    # should convert to UTC
-                    result = Timestamp(result, tz='UTC')
-                    expected_utc = expected - offset * 3600 * 1000000000
-                    self.assertEqual(result.value, expected_utc)
-                    self.assertEqual(tslib.pydt_to_i8(result), expected_utc)
-
-    def test_constructor_with_stringoffset(self):
-        # GH 7833
-        base_str = '2014-07-01 11:00:00+02:00'
-        base_dt = datetime.datetime(2014, 7, 1, 9)
-        base_expected = 1404205200000000000
-
-        # confirm base representation is correct
-        import calendar
-        self.assertEqual(calendar.timegm(base_dt.timetuple()) * 1000000000,
-                         base_expected)
-
-        tests = [(base_str, base_expected),
-                 ('2014-07-01 12:00:00+02:00',
-                  base_expected + 3600 * 1000000000),
-                 ('2014-07-01 11:00:00.000008000+02:00', base_expected + 8000),
-                 ('2014-07-01 11:00:00.000000005+02:00', base_expected + 5)]
-
-        tm._skip_if_no_pytz()
-        tm._skip_if_no_dateutil()
-        import pytz
-        import dateutil
-        timezones = [(None, 0), ('UTC', 0), (pytz.utc, 0), ('Asia/Tokyo', 9),
-                     ('US/Eastern', -4), ('dateutil/US/Pacific', -7),
-                     (pytz.FixedOffset(-180), -3),
-                     (dateutil.tz.tzoffset(None, 18000), 5)]
-
-        for date_str, expected in tests:
-            for result in [Timestamp(date_str)]:
-                # only with timestring
-                self.assertEqual(result.value, expected)
-                self.assertEqual(tslib.pydt_to_i8(result), expected)
-
-                # re-creation shouldn't affect to internal value
-                result = Timestamp(result)
-                self.assertEqual(result.value, expected)
-                self.assertEqual(tslib.pydt_to_i8(result), expected)
-
-            # with timezone
-            for tz, offset in timezones:
-                result = Timestamp(date_str, tz=tz)
-                expected_tz = expected
-                self.assertEqual(result.value, expected_tz)
-                self.assertEqual(tslib.pydt_to_i8(result), expected_tz)
-
-                # should preserve tz
-                result = Timestamp(result)
-                self.assertEqual(result.value, expected_tz)
-                self.assertEqual(tslib.pydt_to_i8(result), expected_tz)
-
-                # should convert to UTC
-                result = Timestamp(result, tz='UTC')
-                expected_utc = expected
-                self.assertEqual(result.value, expected_utc)
-                self.assertEqual(tslib.pydt_to_i8(result), expected_utc)
-
-        # This should be 2013-11-01 05:00 in UTC
-        # converted to Chicago tz
-        result = Timestamp('2013-11-01 00:00:00-0500', tz='America/Chicago')
-        self.assertEqual(result.value, Timestamp('2013-11-01 05:00').value)
-        expected = "Timestamp('2013-11-01 00:00:00-0500', tz='America/Chicago')"  # noqa
-        self.assertEqual(repr(result), expected)
-        self.assertEqual(result, eval(repr(result)))
-
-        # This should be 2013-11-01 05:00 in UTC
-        # converted to Tokyo tz (+09:00)
-        result = Timestamp('2013-11-01 00:00:00-0500', tz='Asia/Tokyo')
-        self.assertEqual(result.value, Timestamp('2013-11-01 05:00').value)
-        expected = "Timestamp('2013-11-01 14:00:00+0900', tz='Asia/Tokyo')"
-        self.assertEqual(repr(result), expected)
-        self.assertEqual(result, eval(repr(result)))
-
-        # GH11708
-        # This should be 2015-11-18 10:00 in UTC
-        # converted to Asia/Katmandu
-        result = Timestamp("2015-11-18 15:45:00+05:45", tz="Asia/Katmandu")
-        self.assertEqual(result.value, Timestamp("2015-11-18 10:00").value)
-        expected = "Timestamp('2015-11-18 15:45:00+0545', tz='Asia/Katmandu')"
-        self.assertEqual(repr(result), expected)
-        self.assertEqual(result, eval(repr(result)))
-
-        # This should be 2015-11-18 10:00 in UTC
-        # converted to Asia/Kolkata
-        result = Timestamp("2015-11-18 15:30:00+05:30", tz="Asia/Kolkata")
-        self.assertEqual(result.value, Timestamp("2015-11-18 10:00").value)
-        expected = "Timestamp('2015-11-18 15:30:00+0530', tz='Asia/Kolkata')"
-        self.assertEqual(repr(result), expected)
-        self.assertEqual(result, eval(repr(result)))
-
-    def test_constructor_invalid(self):
-        with tm.assertRaisesRegexp(TypeError, 'Cannot convert input'):
-            Timestamp(slice(2))
-        with tm.assertRaisesRegexp(ValueError, 'Cannot convert Period'):
-            Timestamp(Period('1000-01-01'))
-
-    def test_constructor_positional(self):
-        # GH 10758
-        with tm.assertRaises(TypeError):
-            Timestamp(2000, 1)
-        with tm.assertRaises(ValueError):
-            Timestamp(2000, 0, 1)
-        with tm.assertRaises(ValueError):
-            Timestamp(2000, 13, 1)
-        with tm.assertRaises(ValueError):
-            Timestamp(2000, 1, 0)
-        with tm.assertRaises(ValueError):
-            Timestamp(2000, 1, 32)
-
-        # GH 11630
-        self.assertEqual(
-            repr(Timestamp(2015, 11, 12)),
-            repr(Timestamp('20151112')))
-
-        self.assertEqual(
-            repr(Timestamp(2015, 11, 12, 1, 2, 3, 999999)),
-            repr(Timestamp('2015-11-12 01:02:03.999999')))
-
-        self.assertIs(Timestamp(None), pd.NaT)
-
-    def test_constructor_keyword(self):
-        # GH 10758
-        with tm.assertRaises(TypeError):
-            Timestamp(year=2000, month=1)
-        with tm.assertRaises(ValueError):
-            Timestamp(year=2000, month=0, day=1)
-        with tm.assertRaises(ValueError):
-            Timestamp(year=2000, month=13, day=1)
-        with tm.assertRaises(ValueError):
-            Timestamp(year=2000, month=1, day=0)
-        with tm.assertRaises(ValueError):
-            Timestamp(year=2000, month=1, day=32)
-
-        self.assertEqual(
-            repr(Timestamp(year=2015, month=11, day=12)),
-            repr(Timestamp('20151112')))
-
-        self.assertEqual(
-            repr(Timestamp(year=2015, month=11, day=12,
-                           hour=1, minute=2, second=3, microsecond=999999)),
-            repr(Timestamp('2015-11-12 01:02:03.999999')))
-
-    def test_constructor_fromordinal(self):
-        base = datetime.datetime(2000, 1, 1)
-
-        ts = Timestamp.fromordinal(base.toordinal(), freq='D')
-        self.assertEqual(base, ts)
-        self.assertEqual(ts.freq, 'D')
-        self.assertEqual(base.toordinal(), ts.toordinal())
-
-        ts = Timestamp.fromordinal(base.toordinal(), tz='US/Eastern')
-        self.assertEqual(pd.Timestamp('2000-01-01', tz='US/Eastern'), ts)
-        self.assertEqual(base.toordinal(), ts.toordinal())
-
-    def test_constructor_offset_depr(self):
-        # GH 12160
-        with tm.assert_produces_warning(FutureWarning,
-                                        check_stacklevel=False):
-            ts = Timestamp('2011-01-01', offset='D')
-        self.assertEqual(ts.freq, 'D')
-
-        with tm.assert_produces_warning(FutureWarning,
-                                        check_stacklevel=False):
-            self.assertEqual(ts.offset, 'D')
-
-        msg = "Can only specify freq or offset, not both"
-        with tm.assertRaisesRegexp(TypeError, msg):
-            Timestamp('2011-01-01', offset='D', freq='D')
-
-    def test_constructor_offset_depr_fromordinal(self):
-        # GH 12160
-        base = datetime.datetime(2000, 1, 1)
-
-        with tm.assert_produces_warning(FutureWarning,
-                                        check_stacklevel=False):
-            ts = Timestamp.fromordinal(base.toordinal(), offset='D')
-        self.assertEqual(pd.Timestamp('2000-01-01'), ts)
-        self.assertEqual(ts.freq, 'D')
-        self.assertEqual(base.toordinal(), ts.toordinal())
-
-        msg = "Can only specify freq or offset, not both"
-        with tm.assertRaisesRegexp(TypeError, msg):
-            Timestamp.fromordinal(base.toordinal(), offset='D', freq='D')
-
-    def test_conversion(self):
-        # GH 9255
-        ts = Timestamp('2000-01-01')
-
-        result = ts.to_pydatetime()
-        expected = datetime.datetime(2000, 1, 1)
-        self.assertEqual(result, expected)
-        self.assertEqual(type(result), type(expected))
-
-        result = ts.to_datetime64()
-        expected = np.datetime64(ts.value, 'ns')
-        self.assertEqual(result, expected)
-        self.assertEqual(type(result), type(expected))
-        self.assertEqual(result.dtype, expected.dtype)
-
-    def test_repr(self):
-        tm._skip_if_no_pytz()
-        tm._skip_if_no_dateutil()
-
-        dates = ['2014-03-07', '2014-01-01 09:00',
-                 '2014-01-01 00:00:00.000000001']
-
-        # dateutil zone change (only matters for repr)
-        import dateutil
-        if (dateutil.__version__ >= LooseVersion('2.3') and
-            (dateutil.__version__ <= LooseVersion('2.4.0') or
-             dateutil.__version__ >= LooseVersion('2.6.0'))):
-            timezones = ['UTC', 'Asia/Tokyo', 'US/Eastern',
-                         'dateutil/US/Pacific']
-        else:
-            timezones = ['UTC', 'Asia/Tokyo', 'US/Eastern',
-                         'dateutil/America/Los_Angeles']
-
-        freqs = ['D', 'M', 'S', 'N']
-
-        for date in dates:
-            for tz in timezones:
-                for freq in freqs:
-
-                    # avoid to match with timezone name
-                    freq_repr = "'{0}'".format(freq)
-                    if tz.startswith('dateutil'):
-                        tz_repr = tz.replace('dateutil', '')
-                    else:
-                        tz_repr = tz
-
-                    date_only = Timestamp(date)
-                    self.assertIn(date, repr(date_only))
-                    self.assertNotIn(tz_repr, repr(date_only))
-                    self.assertNotIn(freq_repr, repr(date_only))
-                    self.assertEqual(date_only, eval(repr(date_only)))
-
-                    date_tz = Timestamp(date, tz=tz)
-                    self.assertIn(date, repr(date_tz))
-                    self.assertIn(tz_repr, repr(date_tz))
-                    self.assertNotIn(freq_repr, repr(date_tz))
-                    self.assertEqual(date_tz, eval(repr(date_tz)))
-
-                    date_freq = Timestamp(date, freq=freq)
-                    self.assertIn(date, repr(date_freq))
-                    self.assertNotIn(tz_repr, repr(date_freq))
-                    self.assertIn(freq_repr, repr(date_freq))
-                    self.assertEqual(date_freq, eval(repr(date_freq)))
-
-                    date_tz_freq = Timestamp(date, tz=tz, freq=freq)
-                    self.assertIn(date, repr(date_tz_freq))
-                    self.assertIn(tz_repr, repr(date_tz_freq))
-                    self.assertIn(freq_repr, repr(date_tz_freq))
-                    self.assertEqual(date_tz_freq, eval(repr(date_tz_freq)))
-
-        # this can cause the tz field to be populated, but it's redundant to
-        # information in the datestring
-        tm._skip_if_no_pytz()
-        import pytz  # noqa
-        date_with_utc_offset = Timestamp('2014-03-13 00:00:00-0400', tz=None)
-        self.assertIn('2014-03-13 00:00:00-0400', repr(date_with_utc_offset))
-        self.assertNotIn('tzoffset', repr(date_with_utc_offset))
-        self.assertIn('pytz.FixedOffset(-240)', repr(date_with_utc_offset))
-        expr = repr(date_with_utc_offset).replace("'pytz.FixedOffset(-240)'",
-                                                  'pytz.FixedOffset(-240)')
-        self.assertEqual(date_with_utc_offset, eval(expr))
-
-    def test_bounds_with_different_units(self):
-        out_of_bounds_dates = ('1677-09-21', '2262-04-12', )
-
-        time_units = ('D', 'h', 'm', 's', 'ms', 'us')
-
-        for date_string in out_of_bounds_dates:
-            for unit in time_units:
-                self.assertRaises(ValueError, Timestamp, np.datetime64(
-                    date_string, dtype='M8[%s]' % unit))
-
-        in_bounds_dates = ('1677-09-23', '2262-04-11', )
-
-        for date_string in in_bounds_dates:
-            for unit in time_units:
-                Timestamp(np.datetime64(date_string, dtype='M8[%s]' % unit))
-
-    def test_tz(self):
-        t = '2014-02-01 09:00'
-        ts = Timestamp(t)
-        local = ts.tz_localize('Asia/Tokyo')
-        self.assertEqual(local.hour, 9)
-        self.assertEqual(local, Timestamp(t, tz='Asia/Tokyo'))
-        conv = local.tz_convert('US/Eastern')
-        self.assertEqual(conv, Timestamp('2014-01-31 19:00', tz='US/Eastern'))
-        self.assertEqual(conv.hour, 19)
-
-        # preserves nanosecond
-        ts = Timestamp(t) + offsets.Nano(5)
-        local = ts.tz_localize('Asia/Tokyo')
-        self.assertEqual(local.hour, 9)
-        self.assertEqual(local.nanosecond, 5)
-        conv = local.tz_convert('US/Eastern')
-        self.assertEqual(conv.nanosecond, 5)
-        self.assertEqual(conv.hour, 19)
-
-    def test_tz_localize_ambiguous(self):
-
-        ts = Timestamp('2014-11-02 01:00')
-        ts_dst = ts.tz_localize('US/Eastern', ambiguous=True)
-        ts_no_dst = ts.tz_localize('US/Eastern', ambiguous=False)
-
-        rng = date_range('2014-11-02', periods=3, freq='H', tz='US/Eastern')
-        self.assertEqual(rng[1], ts_dst)
-        self.assertEqual(rng[2], ts_no_dst)
-        self.assertRaises(ValueError, ts.tz_localize, 'US/Eastern',
-                          ambiguous='infer')
-
-        # GH 8025
-        with tm.assertRaisesRegexp(TypeError,
-                                   'Cannot localize tz-aware Timestamp, use '
-                                   'tz_convert for conversions'):
-            Timestamp('2011-01-01', tz='US/Eastern').tz_localize('Asia/Tokyo')
-
-        with tm.assertRaisesRegexp(TypeError,
-                                   'Cannot convert tz-naive Timestamp, use '
-                                   'tz_localize to localize'):
-            Timestamp('2011-01-01').tz_convert('Asia/Tokyo')
-
-    def test_tz_localize_nonexistent(self):
-        # See issue 13057
-        from pytz.exceptions import NonExistentTimeError
-        times = ['2015-03-08 02:00', '2015-03-08 02:30',
-                 '2015-03-29 02:00', '2015-03-29 02:30']
-        timezones = ['US/Eastern', 'US/Pacific',
-                     'Europe/Paris', 'Europe/Belgrade']
-        for t, tz in zip(times, timezones):
-            ts = Timestamp(t)
-            self.assertRaises(NonExistentTimeError, ts.tz_localize,
-                              tz)
-            self.assertRaises(NonExistentTimeError, ts.tz_localize,
-                              tz, errors='raise')
-            self.assertIs(ts.tz_localize(tz, errors='coerce'),
-                          pd.NaT)
-
-    def test_tz_localize_errors_ambiguous(self):
-        # See issue 13057
-        from pytz.exceptions import AmbiguousTimeError
-        ts = pd.Timestamp('2015-11-1 01:00')
-        self.assertRaises(AmbiguousTimeError,
-                          ts.tz_localize, 'US/Pacific', errors='coerce')
-
-    def test_tz_localize_roundtrip(self):
-        for tz in ['UTC', 'Asia/Tokyo', 'US/Eastern', 'dateutil/US/Pacific']:
-            for t in ['2014-02-01 09:00', '2014-07-08 09:00',
-                      '2014-11-01 17:00', '2014-11-05 00:00']:
-                ts = Timestamp(t)
-                localized = ts.tz_localize(tz)
-                self.assertEqual(localized, Timestamp(t, tz=tz))
-
-                with tm.assertRaises(TypeError):
-                    localized.tz_localize(tz)
-
-                reset = localized.tz_localize(None)
-                self.assertEqual(reset, ts)
-                self.assertTrue(reset.tzinfo is None)
-
-    def test_tz_convert_roundtrip(self):
-        for tz in ['UTC', 'Asia/Tokyo', 'US/Eastern', 'dateutil/US/Pacific']:
-            for t in ['2014-02-01 09:00', '2014-07-08 09:00',
-                      '2014-11-01 17:00', '2014-11-05 00:00']:
-                ts = Timestamp(t, tz='UTC')
-                converted = ts.tz_convert(tz)
-
-                reset = converted.tz_convert(None)
-                self.assertEqual(reset, Timestamp(t))
-                self.assertTrue(reset.tzinfo is None)
-                self.assertEqual(reset,
-                                 converted.tz_convert('UTC').tz_localize(None))
-
-    def test_barely_oob_dts(self):
-        one_us = np.timedelta64(1).astype('timedelta64[us]')
-
-        # By definition we can't go out of bounds in [ns], so we
-        # convert the datetime64s to [us] so we can go out of bounds
-        min_ts_us = np.datetime64(Timestamp.min).astype('M8[us]')
-        max_ts_us = np.datetime64(Timestamp.max).astype('M8[us]')
-
-        # No error for the min/max datetimes
-        Timestamp(min_ts_us)
-        Timestamp(max_ts_us)
-
-        # One us less than the minimum is an error
-        self.assertRaises(ValueError, Timestamp, min_ts_us - one_us)
-
-        # One us more than the maximum is an error
-        self.assertRaises(ValueError, Timestamp, max_ts_us + one_us)
-
-    def test_utc_z_designator(self):
-        self.assertEqual(get_timezone(
-            Timestamp('2014-11-02 01:00Z').tzinfo), 'UTC')
-
-    def test_now(self):
-        # #9000
-        ts_from_string = Timestamp('now')
-        ts_from_method = Timestamp.now()
-        ts_datetime = datetime.datetime.now()
-
-        ts_from_string_tz = Timestamp('now', tz='US/Eastern')
-        ts_from_method_tz = Timestamp.now(tz='US/Eastern')
-
-        # Check that the delta between the times is less than 1s (arbitrarily
-        # small)
-        delta = Timedelta(seconds=1)
-        self.assertTrue(abs(ts_from_method - ts_from_string) < delta)
-        self.assertTrue(abs(ts_datetime - ts_from_method) < delta)
-        self.assertTrue(abs(ts_from_method_tz - ts_from_string_tz) < delta)
-        self.assertTrue(abs(ts_from_string_tz.tz_localize(None) -
-                            ts_from_method_tz.tz_localize(None)) < delta)
-
-    def test_today(self):
-
-        ts_from_string = Timestamp('today')
-        ts_from_method = Timestamp.today()
-        ts_datetime = datetime.datetime.today()
-
-        ts_from_string_tz = Timestamp('today', tz='US/Eastern')
-        ts_from_method_tz = Timestamp.today(tz='US/Eastern')
-
-        # Check that the delta between the times is less than 1s (arbitrarily
-        # small)
-        delta = Timedelta(seconds=1)
-        self.assertTrue(abs(ts_from_method - ts_from_string) < delta)
-        self.assertTrue(abs(ts_datetime - ts_from_method) < delta)
-        self.assertTrue(abs(ts_from_method_tz - ts_from_string_tz) < delta)
-        self.assertTrue(abs(ts_from_string_tz.tz_localize(None) -
-                            ts_from_method_tz.tz_localize(None)) < delta)
-
-    def test_asm8(self):
-        np.random.seed(7960929)
-        ns = [Timestamp.min.value, Timestamp.max.value, 1000, ]
-        for n in ns:
-            self.assertEqual(Timestamp(n).asm8.view('i8'),
-                             np.datetime64(n, 'ns').view('i8'), n)
-        self.assertEqual(Timestamp('nat').asm8.view('i8'),
-                         np.datetime64('nat', 'ns').view('i8'))
-
-    def test_fields(self):
-        def check(value, equal):
-            # that we are int/long like
-            self.assertTrue(isinstance(value, (int, compat.long)))
-            self.assertEqual(value, equal)
-
-        # GH 10050
-        ts = Timestamp('2015-05-10 09:06:03.000100001')
-        check(ts.year, 2015)
-        check(ts.month, 5)
-        check(ts.day, 10)
-        check(ts.hour, 9)
-        check(ts.minute, 6)
-        check(ts.second, 3)
-        self.assertRaises(AttributeError, lambda: ts.millisecond)
-        check(ts.microsecond, 100)
-        check(ts.nanosecond, 1)
-        check(ts.dayofweek, 6)
-        check(ts.quarter, 2)
-        check(ts.dayofyear, 130)
-        check(ts.week, 19)
-        check(ts.daysinmonth, 31)
-        check(ts.daysinmonth, 31)
-
-    def test_nat_fields(self):
-        # GH 10050
-        ts = Timestamp('NaT')
-        self.assertTrue(np.isnan(ts.year))
-        self.assertTrue(np.isnan(ts.month))
-        self.assertTrue(np.isnan(ts.day))
-        self.assertTrue(np.isnan(ts.hour))
-        self.assertTrue(np.isnan(ts.minute))
-        self.assertTrue(np.isnan(ts.second))
-        self.assertTrue(np.isnan(ts.microsecond))
-        self.assertTrue(np.isnan(ts.nanosecond))
-        self.assertTrue(np.isnan(ts.dayofweek))
-        self.assertTrue(np.isnan(ts.quarter))
-        self.assertTrue(np.isnan(ts.dayofyear))
-        self.assertTrue(np.isnan(ts.week))
-        self.assertTrue(np.isnan(ts.daysinmonth))
-        self.assertTrue(np.isnan(ts.days_in_month))
-
-    def test_pprint(self):
-        # GH12622
-        import pprint
-        nested_obj = {'foo': 1,
-                      'bar': [{'w': {'a': Timestamp('2011-01-01')}}] * 10}
-        result = pprint.pformat(nested_obj, width=50)
-        expected = r"""{'bar': [{'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}},
-         {'w': {'a': Timestamp('2011-01-01 00:00:00')}}],
- 'foo': 1}"""
-        self.assertEqual(result, expected)
-
-    def to_datetime_depr(self):
-        # see gh-8254
-        ts = Timestamp('2011-01-01')
-
-        with tm.assert_produces_warning(FutureWarning,
-                                        check_stacklevel=False):
-            expected = datetime.datetime(2011, 1, 1)
-            result = ts.to_datetime()
-            self.assertEqual(result, expected)
-
-    def to_pydatetime_nonzero_nano(self):
-        ts = Timestamp('2011-01-01 9:00:00.123456789')
-
-        # Warn the user of data loss (nanoseconds).
-        with tm.assert_produces_warning(UserWarning,
-                                        check_stacklevel=False):
-            expected = datetime.datetime(2011, 1, 1, 9, 0, 0, 123456)
-            result = ts.to_pydatetime()
-            self.assertEqual(result, expected)
 
 
 class TestDatetimeParsingWrappers(tm.TestCase):
@@ -1117,181 +522,6 @@ class TestArrayToDatetime(tm.TestCase):
             )
 
 
-class TestTimestampNsOperations(tm.TestCase):
-    def setUp(self):
-        self.timestamp = Timestamp(datetime.datetime.utcnow())
-
-    def assert_ns_timedelta(self, modified_timestamp, expected_value):
-        value = self.timestamp.value
-        modified_value = modified_timestamp.value
-
-        self.assertEqual(modified_value - value, expected_value)
-
-    def test_timedelta_ns_arithmetic(self):
-        self.assert_ns_timedelta(self.timestamp + np.timedelta64(-123, 'ns'),
-                                 -123)
-
-    def test_timedelta_ns_based_arithmetic(self):
-        self.assert_ns_timedelta(self.timestamp + np.timedelta64(
-            1234567898, 'ns'), 1234567898)
-
-    def test_timedelta_us_arithmetic(self):
-        self.assert_ns_timedelta(self.timestamp + np.timedelta64(-123, 'us'),
-                                 -123000)
-
-    def test_timedelta_ms_arithmetic(self):
-        time = self.timestamp + np.timedelta64(-123, 'ms')
-        self.assert_ns_timedelta(time, -123000000)
-
-    def test_nanosecond_string_parsing(self):
-        ts = Timestamp('2013-05-01 07:15:45.123456789')
-        # GH 7878
-        expected_repr = '2013-05-01 07:15:45.123456789'
-        expected_value = 1367392545123456789
-        self.assertEqual(ts.value, expected_value)
-        self.assertIn(expected_repr, repr(ts))
-
-        ts = Timestamp('2013-05-01 07:15:45.123456789+09:00', tz='Asia/Tokyo')
-        self.assertEqual(ts.value, expected_value - 9 * 3600 * 1000000000)
-        self.assertIn(expected_repr, repr(ts))
-
-        ts = Timestamp('2013-05-01 07:15:45.123456789', tz='UTC')
-        self.assertEqual(ts.value, expected_value)
-        self.assertIn(expected_repr, repr(ts))
-
-        ts = Timestamp('2013-05-01 07:15:45.123456789', tz='US/Eastern')
-        self.assertEqual(ts.value, expected_value + 4 * 3600 * 1000000000)
-        self.assertIn(expected_repr, repr(ts))
-
-        # GH 10041
-        ts = Timestamp('20130501T071545.123456789')
-        self.assertEqual(ts.value, expected_value)
-        self.assertIn(expected_repr, repr(ts))
-
-    def test_nanosecond_timestamp(self):
-        # GH 7610
-        expected = 1293840000000000005
-        t = Timestamp('2011-01-01') + offsets.Nano(5)
-        self.assertEqual(repr(t), "Timestamp('2011-01-01 00:00:00.000000005')")
-        self.assertEqual(t.value, expected)
-        self.assertEqual(t.nanosecond, 5)
-
-        t = Timestamp(t)
-        self.assertEqual(repr(t), "Timestamp('2011-01-01 00:00:00.000000005')")
-        self.assertEqual(t.value, expected)
-        self.assertEqual(t.nanosecond, 5)
-
-        t = Timestamp(np_datetime64_compat('2011-01-01 00:00:00.000000005Z'))
-        self.assertEqual(repr(t), "Timestamp('2011-01-01 00:00:00.000000005')")
-        self.assertEqual(t.value, expected)
-        self.assertEqual(t.nanosecond, 5)
-
-        expected = 1293840000000000010
-        t = t + offsets.Nano(5)
-        self.assertEqual(repr(t), "Timestamp('2011-01-01 00:00:00.000000010')")
-        self.assertEqual(t.value, expected)
-        self.assertEqual(t.nanosecond, 10)
-
-        t = Timestamp(t)
-        self.assertEqual(repr(t), "Timestamp('2011-01-01 00:00:00.000000010')")
-        self.assertEqual(t.value, expected)
-        self.assertEqual(t.nanosecond, 10)
-
-        t = Timestamp(np_datetime64_compat('2011-01-01 00:00:00.000000010Z'))
-        self.assertEqual(repr(t), "Timestamp('2011-01-01 00:00:00.000000010')")
-        self.assertEqual(t.value, expected)
-        self.assertEqual(t.nanosecond, 10)
-
-    def test_nat_arithmetic(self):
-        # GH 6873
-        i = 2
-        f = 1.5
-
-        for (left, right) in [(pd.NaT, i), (pd.NaT, f), (pd.NaT, np.nan)]:
-            self.assertIs(left / right, pd.NaT)
-            self.assertIs(left * right, pd.NaT)
-            self.assertIs(right * left, pd.NaT)
-            with tm.assertRaises(TypeError):
-                right / left
-
-        # Timestamp / datetime
-        t = Timestamp('2014-01-01')
-        dt = datetime.datetime(2014, 1, 1)
-        for (left, right) in [(pd.NaT, pd.NaT), (pd.NaT, t), (pd.NaT, dt)]:
-            # NaT __add__ or __sub__ Timestamp-like (or inverse) returns NaT
-            self.assertIs(right + left, pd.NaT)
-            self.assertIs(left + right, pd.NaT)
-            self.assertIs(left - right, pd.NaT)
-            self.assertIs(right - left, pd.NaT)
-
-        # timedelta-like
-        # offsets are tested in test_offsets.py
-
-        delta = datetime.timedelta(3600)
-        td = Timedelta('5s')
-
-        for (left, right) in [(pd.NaT, delta), (pd.NaT, td)]:
-            # NaT + timedelta-like returns NaT
-            self.assertIs(right + left, pd.NaT)
-            self.assertIs(left + right, pd.NaT)
-            self.assertIs(right - left, pd.NaT)
-            self.assertIs(left - right, pd.NaT)
-
-        # GH 11718
-        tm._skip_if_no_pytz()
-        import pytz
-
-        t_utc = Timestamp('2014-01-01', tz='UTC')
-        t_tz = Timestamp('2014-01-01', tz='US/Eastern')
-        dt_tz = pytz.timezone('Asia/Tokyo').localize(dt)
-
-        for (left, right) in [(pd.NaT, t_utc), (pd.NaT, t_tz),
-                              (pd.NaT, dt_tz)]:
-            # NaT __add__ or __sub__ Timestamp-like (or inverse) returns NaT
-            self.assertIs(right + left, pd.NaT)
-            self.assertIs(left + right, pd.NaT)
-            self.assertIs(left - right, pd.NaT)
-            self.assertIs(right - left, pd.NaT)
-
-        # int addition / subtraction
-        for (left, right) in [(pd.NaT, 2), (pd.NaT, 0), (pd.NaT, -3)]:
-            self.assertIs(right + left, pd.NaT)
-            self.assertIs(left + right, pd.NaT)
-            self.assertIs(left - right, pd.NaT)
-            self.assertIs(right - left, pd.NaT)
-
-    def test_nat_arithmetic_index(self):
-        # GH 11718
-
-        # datetime
-        tm._skip_if_no_pytz()
-
-        dti = pd.DatetimeIndex(['2011-01-01', '2011-01-02'], name='x')
-        exp = pd.DatetimeIndex([pd.NaT, pd.NaT], name='x')
-        self.assert_index_equal(dti + pd.NaT, exp)
-        self.assert_index_equal(pd.NaT + dti, exp)
-
-        dti_tz = pd.DatetimeIndex(['2011-01-01', '2011-01-02'],
-                                  tz='US/Eastern', name='x')
-        exp = pd.DatetimeIndex([pd.NaT, pd.NaT], name='x', tz='US/Eastern')
-        self.assert_index_equal(dti_tz + pd.NaT, exp)
-        self.assert_index_equal(pd.NaT + dti_tz, exp)
-
-        exp = pd.TimedeltaIndex([pd.NaT, pd.NaT], name='x')
-        for (left, right) in [(pd.NaT, dti), (pd.NaT, dti_tz)]:
-            self.assert_index_equal(left - right, exp)
-            self.assert_index_equal(right - left, exp)
-
-        # timedelta
-        tdi = pd.TimedeltaIndex(['1 day', '2 day'], name='x')
-        exp = pd.DatetimeIndex([pd.NaT, pd.NaT], name='x')
-        for (left, right) in [(pd.NaT, tdi)]:
-            self.assert_index_equal(left + right, exp)
-            self.assert_index_equal(right + left, exp)
-            self.assert_index_equal(left - right, exp)
-            self.assert_index_equal(right - left, exp)
-
-
 class TestTslib(tm.TestCase):
     def test_intraday_conversion_factors(self):
         self.assertEqual(period_asfreq(
@@ -1459,86 +689,6 @@ class TestTslib(tm.TestCase):
         msg = pd.tseries.frequencies._INVALID_FREQ_ERROR
         with self.assertRaisesRegexp(ValueError, msg):
             stamp.round('foo')
-
-
-class TestTimestampOps(tm.TestCase):
-    def test_timestamp_and_datetime(self):
-        self.assertEqual((Timestamp(datetime.datetime(
-            2013, 10, 13)) - datetime.datetime(2013, 10, 12)).days, 1)
-        self.assertEqual((datetime.datetime(2013, 10, 12) -
-                          Timestamp(datetime.datetime(2013, 10, 13))).days, -1)
-
-    def test_timestamp_and_series(self):
-        timestamp_series = Series(date_range('2014-03-17', periods=2, freq='D',
-                                             tz='US/Eastern'))
-        first_timestamp = timestamp_series[0]
-
-        delta_series = Series([np.timedelta64(0, 'D'), np.timedelta64(1, 'D')])
-        assert_series_equal(timestamp_series - first_timestamp, delta_series)
-        assert_series_equal(first_timestamp - timestamp_series, -delta_series)
-
-    def test_addition_subtraction_types(self):
-        # Assert on the types resulting from Timestamp +/- various date/time
-        # objects
-        datetime_instance = datetime.datetime(2014, 3, 4)
-        timedelta_instance = datetime.timedelta(seconds=1)
-        # build a timestamp with a frequency, since then it supports
-        # addition/subtraction of integers
-        timestamp_instance = date_range(datetime_instance, periods=1,
-                                        freq='D')[0]
-
-        self.assertEqual(type(timestamp_instance + 1), Timestamp)
-        self.assertEqual(type(timestamp_instance - 1), Timestamp)
-
-        # Timestamp + datetime not supported, though subtraction is supported
-        # and yields timedelta more tests in tseries/base/tests/test_base.py
-        self.assertEqual(
-            type(timestamp_instance - datetime_instance), Timedelta)
-        self.assertEqual(
-            type(timestamp_instance + timedelta_instance), Timestamp)
-        self.assertEqual(
-            type(timestamp_instance - timedelta_instance), Timestamp)
-
-        # Timestamp +/- datetime64 not supported, so not tested (could possibly
-        # assert error raised?)
-        timedelta64_instance = np.timedelta64(1, 'D')
-        self.assertEqual(
-            type(timestamp_instance + timedelta64_instance), Timestamp)
-        self.assertEqual(
-            type(timestamp_instance - timedelta64_instance), Timestamp)
-
-    def test_addition_subtraction_preserve_frequency(self):
-        timestamp_instance = date_range('2014-03-05', periods=1, freq='D')[0]
-        timedelta_instance = datetime.timedelta(days=1)
-        original_freq = timestamp_instance.freq
-        self.assertEqual((timestamp_instance + 1).freq, original_freq)
-        self.assertEqual((timestamp_instance - 1).freq, original_freq)
-        self.assertEqual(
-            (timestamp_instance + timedelta_instance).freq, original_freq)
-        self.assertEqual(
-            (timestamp_instance - timedelta_instance).freq, original_freq)
-
-        timedelta64_instance = np.timedelta64(1, 'D')
-        self.assertEqual(
-            (timestamp_instance + timedelta64_instance).freq, original_freq)
-        self.assertEqual(
-            (timestamp_instance - timedelta64_instance).freq, original_freq)
-
-    def test_resolution(self):
-
-        for freq, expected in zip(['A', 'Q', 'M', 'D', 'H', 'T',
-                                   'S', 'L', 'U'],
-                                  [RESO_DAY, RESO_DAY,
-                                   RESO_DAY, RESO_DAY,
-                                   RESO_HR, RESO_MIN,
-                                   RESO_SEC, RESO_MS,
-                                   RESO_US]):
-            for tz in [None, 'Asia/Tokyo', 'US/Eastern',
-                       'dateutil/US/Eastern']:
-                idx = date_range(start='2013-04-01', periods=30, freq=freq,
-                                 tz=tz)
-                result = period.resolution(idx.asi8, idx.tz)
-                self.assertEqual(result, expected)
 
 
 if __name__ == '__main__':
